+++ tests/test_visitors/test_ast/test_classes/test_class_attributes/test_shadow_attributes.py
@@ -48,6 +48,15 @@ class ClassWithAttrs:
         self.{1} = 2
 """
 
+class_with_unrelated_decorator = """
+@not_a_dataclass
+class ClassWithAttrs:
+    {0}: int = 0
+
+    def __post_init__(self) -> None:
+        self.{1} = 2
+"""
+
 # Safe:
 
 class_annotation = """
@@ -98,6 +107,15 @@ class ClassWithAttrs:
         self.{1} = 2
 """
 
+class_with_dataclass_decorator = """
+@dataclass
+class ClassWithAttrs:
+    {0}: int = 0
+
+    def __post_init__(self) -> None:
+        self.{1} = 2
+"""
+
 regular_assigns = """
 {0} = 0
 {1} = 2
@@ -112,6 +130,7 @@ regular_assigns = """
         class_attribute_runtime,
         class_attribute_annotated,
         class_attribute_logic,
+        class_with_unrelated_decorator,
     ],
 )
 @pytest.mark.parametrize(
@@ -154,6 +173,8 @@ def test_incorrect_fields(
         class_attribute_logic,
         class_attribute_regular_assign,
         class_attribute_with_other,
+        class_with_unrelated_decorator,
+        class_with_dataclass_decorator,
         regular_assigns,
     ],
 )
@@ -191,6 +212,7 @@ def test_correct_fields(
         regular_assigns,
         class_complex_attribute,
         class_complex_attribute_annotated,
+        class_with_dataclass_decorator,
     ],
 )
 @pytest.mark.parametrize(